requirements.txt
numpy
scipy
scikit-learn
scikit-image
Pillow
cython
matplotlib
scikit-image
#tensorflow-gpu>=1.3.0
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
